refactoring: - removed unnecessary components
