+++ src/mock/comments.js
@@ -1,8 +1,9 @@
 import { authors, commentContent, datesOfComment, smiles } from './data';
 import { generateContent } from './random-content';
+import {nanoid} from 'nanoid';
 
 export const generateComment = () => ({
-  id: 1,
+  id: nanoid(),
   author: generateContent(authors),
   commentMessage: generateContent(commentContent),
   emotion: generateContent(smiles),
