+++ src/presenter/film-details-presenter.js
@@ -1,23 +1,13 @@
 import FilmDetailsView from '../view/film-details/film-details-view';
-import FilmDetailsFormView from '../view/film-details/film-details-form-view';
-import FilmDetailsTopContainerView from '../view/film-details/film-details-top-container-view';
-import FilmDetailsCloseView from '../view/film-details/film-details-close-view';
-import FilmDetailsInfoView from '../view/film-details/film-details-info-view';
 import FilmDetailsControlsView from '../view/film-details/film-details-controls-view';
-import FilmDetailsBottomContainerView from '../view/film-details/film-details-bottom-container-view';
-import {generateComment} from '../mock/comments';
 import {remove, render, RenderPosition} from '../utils/render';
 
 export default class FilmDetailsPresenter {
   #filmDetailsComponent = null;
-  #filmDetailsFormComponent = null;
-  #filmDetailsTopContainerComponent = null;
-  #filmDetailsCloseComponent = null;
-  #filmDetailsInfoComponent = null;
   #filmDetailsControlsComponent = null;
-  #filmDetailsBottomContainerComponent = null;
 
   #footer = document.querySelector('.footer');
+  #filmDetailsTopContainer = null;
 
   #state = null;
   #changeData = null;
@@ -31,14 +21,10 @@ export default class FilmDetailsPresenter {
   init = (film) => {
     this.#film = film;
 
-    this.#filmDetailsComponent = new FilmDetailsView();
-    this.#filmDetailsFormComponent = new FilmDetailsFormView();
-    this.#filmDetailsTopContainerComponent = new FilmDetailsTopContainerView();
-    this.#filmDetailsCloseComponent = new FilmDetailsCloseView();
-    this.#filmDetailsInfoComponent = new FilmDetailsInfoView(this.#film);
+    this.#filmDetailsComponent = new FilmDetailsView(this.#film);
     this.#filmDetailsControlsComponent = new FilmDetailsControlsView(this.#film);
-    this.#filmDetailsBottomContainerComponent = new FilmDetailsBottomContainerView(this.#film, generateComment);
-    this.#filmDetailsCloseComponent.setClosePopupClickHandler(this.#handleClosePopup);
+    this.#filmDetailsTopContainer = this.#filmDetailsComponent.element.querySelector('.film-details__top-container');
+    this.#filmDetailsComponent.setClosePopupClickHandler(this.#handleClosePopup);
     this.#setControlsHandlers();
 
     this.#renderPopup();
@@ -46,12 +32,7 @@ export default class FilmDetailsPresenter {
 
   #renderPopup = () => {
     render(this.#footer, this.#filmDetailsComponent, RenderPosition.BEFOREEND);
-    render(this.#filmDetailsComponent, this.#filmDetailsFormComponent, RenderPosition.BEFOREEND);
-    render(this.#filmDetailsFormComponent, this.#filmDetailsTopContainerComponent, RenderPosition.BEFOREEND);
-    render(this.#filmDetailsFormComponent, this.#filmDetailsBottomContainerComponent, RenderPosition.BEFOREEND);
-    render(this.#filmDetailsTopContainerComponent, this.#filmDetailsCloseComponent, RenderPosition.BEFOREEND);
-    render(this.#filmDetailsTopContainerComponent, this.#filmDetailsInfoComponent, RenderPosition.BEFOREEND);
-    render(this.#filmDetailsTopContainerComponent, this.#filmDetailsControlsComponent, RenderPosition.BEFOREEND);
+    render(this.#filmDetailsTopContainer, this.#filmDetailsControlsComponent, RenderPosition.BEFOREEND);
   }
 
   #handleFavoriteClick = () => {
@@ -93,6 +74,6 @@ export default class FilmDetailsPresenter {
     remove(this.#filmDetailsControlsComponent);
     this.#filmDetailsControlsComponent = new FilmDetailsControlsView(film);
     this.#setControlsHandlers();
-    render(this.#filmDetailsTopContainerComponent, this.#filmDetailsControlsComponent, RenderPosition.BEFOREEND);
+    render(this.#filmDetailsTopContainer, this.#filmDetailsControlsComponent, RenderPosition.BEFOREEND);
   }
 }
