+++ src/view/film-details/film-details-view.js
@@ -1,20 +1,143 @@
 
 import ComponentView from '../component-view';
+import dayjs from 'dayjs';
+import {getTimeFromMins} from '../../utils/common';
 
-const createFilmDetailsTemplate = () => (
-  `<section class="film-details">
+const createFilmDetailsTemplate = (film) => {
+  const {name, description, poster, rating, release, director, actor, writers, runtime, genre} = film;
+  const fullDate = dayjs(release.date).format('D MMMM YYYY');
 
+  return (
+    `<section class="film-details">
+      <form class="film-details__inner" action="" method="get">
+        <div class="film-details__top-container">
+          <div class="film-details__close">
+            <button class="film-details__close-btn" type="button">close</button>
+          </div>
+
+          <div class="film-details__info-wrap">
+            <div class="film-details__poster">
+              <img class="film-details__poster-img" src="./images/posters/${poster}" alt="">
+
+              <p class="film-details__age">18+</p>
+            </div>
+
+            <div class="film-details__info">
+              <div class="film-details__info-head">
+                <div class="film-details__title-wrap">
+                  <h3 class="film-details__title">${name ? name : ''}</h3>
+                  <p class="film-details__title-original">Original: ${name ? name : ''}</p>
+                </div>
+
+                <div class="film-details__rating">
+                  <p class="film-details__total-rating">${rating ? rating : ''}</p>
+                </div>
+              </div>
+
+              <table class="film-details__table">
+                <tr class="film-details__row">
+                  <td class="film-details__term">Director</td>
+                  <td class="film-details__cell">${director ? director : ''}</td>
+                </tr>
+                <tr class="film-details__row">
+                  <td class="film-details__term">Writers</td>
+                  <td class="film-details__cell">${writers ? writers.join(', ') : ''}</td>
+                </tr>
+                <tr class="film-details__row">
+                  <td class="film-details__term">Actors</td>
+                  <td class="film-details__cell">${actor ? actor.join(', ') : ''}</td>
+                </tr>
+                <tr class="film-details__row">
+                  <td class="film-details__term">Release Date</td>
+                  <td class="film-details__cell">${fullDate}</td>
+                </tr>
+                <tr class="film-details__row">
+                  <td class="film-details__term">Runtime</td>
+                  <td class="film-details__cell">${runtime ? getTimeFromMins(runtime) : ''}</td>
+                </tr>
+                <tr class="film-details__row">
+                  <td class="film-details__term">Country</td>
+                  <td class="film-details__cell">${release.releaseCountry}</td>
+                </tr>
+                <tr class="film-details__row">
+                  <td class="film-details__term">Genres</td>
+                  <td class="film-details__cell">
+                    ${genre.map((genreElement) => `<span class="film-details__genre">${genreElement}</span>`).join('')}
+                  </td>
+                </tr>
+              </table>
+
+              <p class="film-details__film-description">
+                ${description ? description : ''}
+              </p>
+          </div>
+        </div>
+
+        </div>
+
+        <div class="film-details__bottom-container">
+          <section class="film-details__comments-wrap">
+            <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${film.comments.length}</span></h3>
+
+            <ul class="film-details__comments-list">
+            </ul>
+
+            <div class="film-details__new-comment">
+              <div class="film-details__add-emoji-label"></div>
+
+              <label class="film-details__comment-label">
+                <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
+              </label>
+
+              <div class="film-details__emoji-list">
+                <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
+                <label class="film-details__emoji-label" for="emoji-smile">
+                  <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
+                </label>
+
+                <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
+                <label class="film-details__emoji-label" for="emoji-sleeping">
+                  <img src="./images/emoji/sleeping.png" width="30" height="30" alt="emoji">
+                </label>
+
+                <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
+                <label class="film-details__emoji-label" for="emoji-puke">
+                  <img src="./images/emoji/puke.png" width="30" height="30" alt="emoji">
+                </label>
+
+                <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
+                <label class="film-details__emoji-label" for="emoji-angry">
+                  <img src="./images/emoji/angry.png" width="30" height="30" alt="emoji">
+                </label>
+              </div>
+            </div>
+          </section>
+        </div>
+      </form>
     </section>`
-);
+  );
+};
 
 
 export default class FilmDetailsView extends ComponentView {
+  #film = null;
 
-  constructor() {
+  constructor(film) {
     super();
+    this.#film = film;
   }
 
   get template() {
-    return createFilmDetailsTemplate();
+    return createFilmDetailsTemplate(this.#film);
+  }
+
+  setClosePopupClickHandler = (callback) => {
+    this._callback.closePopup = callback;
+    this.element.querySelector('.film-details__close-btn').addEventListener('click', this.#closePopupClickHandler);
+  }
+
+  #closePopupClickHandler = (e) => {
+    e.preventDefault();
+    this._callback.closePopup();
   }
 }
